refactor(ProfileForm): restructure component for improved state management and modularity
- Refactored ProfileForm component to utilize custom hooks for state and data management.
- Introduced useProfileFormState and useProfileData hooks to encapsulate form state logic and GraphQL operations.
- Created sub-components: PersonalDetailsCard, ContactInfoCard, and ActivityCards for better separation of concerns.
- Updated utility functions for handling profile updates and localStorage interactions.
- Enhanced code readability and maintainability by reducing complexity in the main component.
