+++ src/types/ProfileForm/type.ts
@@ -1,0 +1,23 @@
+export interface FormStateType {
+  addressLine1: string;
+  addressLine2: string;
+  birthDate: null | string;
+  emailAddress: string;
+  city: string;
+  avatar: File | null;
+  avatarURL: string;
+  countryCode: string;
+  description: string;
+  educationGrade: string;
+  employmentStatus: string;
+  homePhoneNumber: string;
+  maritalStatus: string;
+  mobilePhoneNumber: string;
+  name: string;
+  natalSex: string;
+  naturalLanguageCode: string;
+  password: string;
+  postalCode: string;
+  state: string;
+  workPhoneNumber: string;
+}
